Fix handleTssMismatches crashes. (#12330)
handleTssMismatches(DatabaseContext* cx) uses a pointer to DatabaseContext
object, which can be destroyed when "tr" is reset within this actor. However,
the actor can't be destroyed because it's on the stack. Introducing this delay
gives a chance to cancel the actor.

(cherry picked from commit d4370b909cd31661bbc8b8c1b803563031b822e6)